moved core bot files into subfolder so index is easier to parse
